--- gda/templates/gda/vote.html
@@ -1,36 +1,0 @@
-{% extends "stalkeador/header.html" %}
-
-{% block content %}
-
-    <h4 class='text-center'>This is a vote page</h4>
-
-    {% if token %}
-        <h3 class='text-center'>{{ token.student }}</h3>
-    {% else %}
-        <h4 class='text-center'>Theres's no valid token</h4>
-    {% endif %}
-
-    {% if discipline %}
-    <h3 class="text-center"> {{ discipline }} </h3>
-    <table>
-        <caption>Informações da Disciplina</caption>
-        <thead>
-            <tr>
-                <th width="150"></th>
-                <th width="150"></th>
-            </tr>
-        </thead>
-        <tbody>
-            <tr>
-                <td>
-                    Código - Turma
-                </td>
-                <td>
-                    {{ discipline.code }} - {{ discipline.offering_id }}
-                </td>
-            </tr>
-        </tbody>
-    </table>
-    {% endif %}
-
-{% endblock %}
